requirements.txt
aiogram==3.0.0b7
aiohttp==3.8.6
pymongo
python-dotenv
motor
telethon
